+++ src/quickjs_utils/modules.rs
@@ -1,14 +1,19 @@
+use crate::eserror::EsError;
 use crate::esscript::EsScript;
+use crate::quickjs_utils::atoms;
+use crate::quickjs_utils::atoms::JSAtomRef;
 use crate::quickjscontext::QuickJsContext;
 use crate::quickjsruntime::QuickJsRuntime;
+use crate::valueref::JSValueRef;
 use libquickjs_sys as q;
 use log::trace;
 use std::cell::RefCell;
 use std::collections::HashSet;
 use std::ffi::{CStr, CString};
+use std::os::raw::c_int;
 
 thread_local! {
-    // todo refactor to per ctx / use lookup method
+    // todo refactor to per ctx (map should be instance of quickjscontext instead of thread_local)
     static LOADED_MODULE_REGISTRY: RefCell<HashSet<String>> = RefCell::new(HashSet::new());
 }
 
@@ -30,6 +35,58 @@ pub fn detect_module(source: &str) -> bool {
         .ok()
         .expect("could not create CString to to null term in source");
     unsafe { q::JS_DetectModule(cstr.as_ptr(), source.len() as u64) != 0 }
+}
+
+/// create new Module (JSModuleDef struct) which can be populated with exports after (and from) the init_func
+pub unsafe fn new_module(
+    ctx: *mut q::JSContext,
+    name: &str,
+    init_func: q::JSModuleInitFunc,
+) -> Result<*mut q::JSModuleDef, EsError> {
+    let name_cstr = CString::new(name).map_err(|_e| EsError::new_str("CString failed"))?;
+    Ok(q::JS_NewCModule(ctx, name_cstr.as_ptr(), init_func))
+}
+
+/// set an export in a JSModuleDef, this should be called AFTER the init_func(as passed to new_module()) is called
+/// please note that you always need to use this in combination with add_module_export()
+pub unsafe fn set_module_export(
+    ctx: *mut q::JSContext,
+    module: *mut q::JSModuleDef,
+    export_name: &str,
+    js_val: JSValueRef,
+) -> Result<(), EsError> {
+    let name_cstr = CString::new(export_name).map_err(|_e| EsError::new_str("CString failed"))?;
+    let res = q::JS_SetModuleExport(ctx, module, name_cstr.as_ptr(), *js_val.borrow_value());
+    if res == 0 {
+        Ok(())
+    } else {
+        Err(EsError::new_str("JS_SetModuleExport failed"))
+    }
+}
+
+/// set an export in a JSModuleDef, this should be called BEFORE this init_func(as passed to new_module()) is called
+pub unsafe fn add_module_export(
+    ctx: *mut q::JSContext,
+    module: *mut q::JSModuleDef,
+    export_name: &str,
+) -> Result<(), EsError> {
+    let name_cstr = CString::new(export_name).map_err(|_e| EsError::new_str("CString failed"))?;
+    let res = q::JS_AddModuleExport(ctx, module, name_cstr.as_ptr());
+    if res == 0 {
+        Ok(())
+    } else {
+        Err(EsError::new_str("JS_SetModuleExport failed"))
+    }
+}
+
+/// get the name of an JSModuleDef struct
+pub unsafe fn get_module_name(
+    ctx: *mut q::JSContext,
+    module: *mut q::JSModuleDef,
+) -> Result<String, EsError> {
+    let atom_raw = q::JS_GetModuleName(ctx, module);
+    let atom_ref = JSAtomRef::new(ctx, atom_raw);
+    atoms::to_string(ctx, &atom_ref)
 }
 
 unsafe extern "C" fn js_module_normalize(
@@ -119,26 +176,65 @@ unsafe extern "C" fn js_module_normalize(
     }
 }
 
+unsafe extern "C" fn native_module_init(
+    ctx: *mut q::JSContext,
+    module: *mut q::JSModuleDef,
+) -> c_int {
+    let module_name = get_module_name(ctx, module)
+        .ok()
+        .expect("could not get name");
+    log::trace!("native_module_init: {}", module_name);
+
+    QuickJsRuntime::do_with(|q_js_rt| {
+        if let Some(module_loader) = &q_js_rt.native_module_loader {
+            QuickJsContext::with_context(ctx, |q_ctx| {
+                for (name, val) in module_loader.get_module_exports(q_ctx, module_name.as_str()) {
+                    set_module_export(ctx, module, name.as_str(), val)
+                        .ok()
+                        .expect("could not set export");
+                }
+            })
+        }
+    });
+
+    0 // ok
+}
+
 unsafe extern "C" fn js_module_loader(
-    _ctx: *mut q::JSContext,
-    module_name: *const ::std::os::raw::c_char,
+    ctx: *mut q::JSContext,
+    module_name_raw: *const ::std::os::raw::c_char,
     _opaque: *mut ::std::os::raw::c_void,
 ) -> *mut q::JSModuleDef {
     log::trace!("js_module_loader called.");
 
-    let module_name_c = CStr::from_ptr(module_name);
-    let res = module_name_c.to_str();
+    let module_name_c = CStr::from_ptr(module_name_raw);
+    let module_name = module_name_c
+        .to_str()
+        .ok()
+        .expect("could not get module name");
 
-    if res.is_err() {
-        panic!("failed to get module name: {}", res.err().unwrap());
-    }
+    log::trace!("js_module_loader called: {}", module_name);
 
-    log::trace!(
-        "js_module_loader called: {}",
-        res.expect("could not get module_name")
-    );
+    QuickJsRuntime::do_with(|q_js_rt| {
+        if let Some(module_loader) = &q_js_rt.native_module_loader {
+            QuickJsContext::with_context(ctx, |q_ctx| {
+                let module = new_module(ctx, module_name, Some(native_module_init))
+                    .ok()
+                    .expect("could not create new module");
 
-    std::ptr::null_mut()
+                for name in module_loader.get_module_export_names(q_ctx, module_name) {
+                    add_module_export(ctx, module, name.as_str())
+                        .ok()
+                        .expect("could not add export");
+                }
+
+                //std::ptr::null_mut()
+                module
+            })
+        } else {
+            std::ptr::null_mut()
+        }
+    })
 }
 
 #[cfg(test)]
@@ -148,6 +244,21 @@ pub mod tests {
     use crate::quickjs_utils::modules::detect_module;
     use std::sync::Arc;
     use std::time::Duration;
+
+    #[test]
+    fn test_native_modules() {
+        let rt: Arc<EsRuntime> = crate::esruntime::tests::TEST_ESRT.clone();
+        let mres = rt.eval_module_sync(EsScript::new(
+            "test.mes",
+            "import {a, b, c} from 'greco://testmodule1';\nconsole.log('testmodule1.a = %s, testmodule1.b = %s, testmodule1.c = %s', a, b, c);",
+        ));
+        match mres {
+            Ok(_module_res) => {}
+            Err(e) => {
+                panic!("test_native_modules failed: {}", e)
+            }
+        }
+    }
 
     #[test]
     fn test_detect() {